pyproject.toml
[project]
name = "vector-institute-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"google-search-results>=2.4.2",
	"langchain>=1.1.0",
	"langchain-openai>=1.1.0",
	"python-dotenv>=1.2.1",
	"SQLAlchemy>=2.0.0",
]
